feat: use float prices instead of int, fix: cancel editing order now works as intended

+++ backend/database/models.py
@@ -36,14 +36,14 @@ class TokenData(BaseModel):
 
 class IngredientModel(BaseModel):
     name: str
-    price: int = Field(gt=0, description="The price must be greater than zero")
+    price: float = Field(gt=0, description="The price must be greater than zero")
     weight: int = Field(gt=0, description="The weight must be greater than zero")
 
 
 class BurgerModel(BaseModel):
     name: str
     ingredients: list[IngredientModel]
-    price: int = Field(gt=0, description="The price must be greater than zero")
+    price: float = Field(gt=0, description="The price must be greater than zero")
     weight: int = Field(gt=0, description="The weight must be greater than zero")
 
 
@@ -51,7 +51,7 @@ class OrderModel(BaseModel):
     customer: str
     status: str
     content: list[BurgerModel]
-    price: int = Field(gt=0, description="The price must be greater than zero")
+    price: float = Field(gt=0, description="The price must be greater than zero")
     weight: int = Field(gt=0, description="The weight must be greater than zero")
     creation_datetime: str
